move prover to a different crate
