+++ crates/graph/Cargo.toml
@@ -16,6 +16,8 @@ num-traits.workspace = true
 serde.workspace = true
 typetag.workspace = true
 luminair_air = { path = "../air" }
+luminair_utils = { path = "../utils" }
+luminair_prover = { path = "../prover" }
 itertools = "0.14.0"
 
 [dev-dependencies]
