+++ crates/graph/src/graph.rs
@@ -3,59 +3,32 @@ use crate::{
         prim::{CopyFromStwo, CopyToStwo, LuminairConstant},
         HasProcessTrace,
     },
-    settings::CircuitSettings,
     utils::compute_padded_range_from_srcs,
 };
 use luminair_air::{
     components::{
-        add::{
-            self,
-            table::{AddColumn, AddTraceTable},
-        },
+        add::table::{AddColumn, AddTraceTable},
         lookups::{
-            self,
             sin::{table::SinLookupTraceTable, SinLookup},
             Lookups,
         },
-        max_reduce::{
-            self,
-            table::{MaxReduceColumn, MaxReduceTraceTable},
-        },
-        mul::{
-            self,
-            table::{MulColumn, MulTraceTable},
-        },
-        recip::{
-            self,
-            table::{RecipColumn, RecipTraceTable},
-        },
-        sin::{
-            self,
-            table::{SinColumn, SinTraceTable},
-        },
-        sqrt::{
-            self,
-            table::{SqrtColumn, SqrtTraceTable},
-        },
-        sum_reduce::{
-            self,
-            table::{SumReduceColumn, SumReduceTraceTable},
-        },
-        LuminairComponents, LuminairInteractionElements, TraceError,
+        max_reduce::table::{MaxReduceColumn, MaxReduceTraceTable},
+        mul::table::{MulColumn, MulTraceTable},
+        recip::table::{RecipColumn, RecipTraceTable},
+        sin::table::{SinColumn, SinTraceTable},
+        sqrt::table::{SqrtColumn, SqrtTraceTable},
+        sum_reduce::table::{SumReduceColumn, SumReduceTraceTable},
+        LuminairComponents, LuminairInteractionElements,
     },
     pie::{
         ExecutionResources, InputInfo, LuminairPie, NodeInfo, OpCounter, OutputInfo, TraceTable,
     },
-    preprocessed::{
-        lookups_to_preprocessed_column,
-        LookupLayout,
-        PreProcessedTrace,
-        Range,
-        SinPreProcessed, // SinLUT
-    },
+    preprocessed::{lookups_to_preprocessed_column, LookupLayout, PreProcessedTrace, Range},
+    settings::CircuitSettings,
     utils::{calculate_log_size, log_sum_valid},
-    LuminairClaim, LuminairInteractionClaim, LuminairInteractionClaimGenerator, LuminairProof,
+    LuminairProof,
 };
+use luminair_utils::LuminairError;
 use luminal::{
     op::*,
     prelude::{petgraph::visit::EdgeRef, *},
@@ -64,37 +37,12 @@ use numerair::Fixed;
 use stwo_prover::{
     constraint_framework::{INTERACTION_TRACE_IDX, ORIGINAL_TRACE_IDX, PREPROCESSED_TRACE_IDX},
     core::{
-        backend::simd::SimdBackend,
         channel::Blake2sChannel,
-        pcs::{CommitmentSchemeProver, CommitmentSchemeVerifier, PcsConfig},
-        poly::circle::{CanonicCoset, PolyOps},
-        prover::{self, verify, ProvingError, VerificationError},
+        pcs::{CommitmentSchemeVerifier, PcsConfig},
+        prover::verify,
         vcs::blake2_merkle::{Blake2sMerkleChannel, Blake2sMerkleHasher},
     },
 };
-use thiserror::Error;
-
-/// Errors that can occur during LuminAIR graph processing, proof generation, or verification.
-#[derive(Clone, Debug, Error)]
-pub enum LuminairError {
-    #[error(transparent)]
-    TraceError(#[from] TraceError),
-
-    #[error("Main trace generation failed.")]
-    MainTraceEvalGenError,
-
-    #[error("Interaction trace generation failed.")]
-    InteractionTraceEvalGenError,
-
-    #[error(transparent)]
-    ProverError(#[from] ProvingError),
-
-    #[error(transparent)]
-    StwoVerifierError(#[from] VerificationError),
-
-    #[error("{0} LogUp values do not match.")]
-    InvalidLogUp(String),
-}
 
 /// Trait defining the core functionality of a LuminAIR computation graph.
 ///
@@ -105,14 +53,7 @@ pub trait LuminairGraph {
     fn gen_circuit_settings(&mut self) -> CircuitSettings;
 
     /// Generates an execution trace for the graph's computation.
-    fn gen_trace(&mut self, settings: &mut CircuitSettings) -> Result<LuminairPie, TraceError>;
-
-    /// Generates a proof of the graph's execution using the provided trace.
-    fn prove(
-        &mut self,
-        pie: LuminairPie,
-        settings: CircuitSettings,
-    ) -> Result<LuminairProof<Blake2sMerkleHasher>, LuminairError>;
+    fn gen_trace(&mut self, settings: &mut CircuitSettings) -> Result<LuminairPie, LuminairError>;
 
     /// Verifies a proof to ensure integrity of graph's computation.
     fn verify(
@@ -193,7 +134,7 @@ impl LuminairGraph for Graph {
     /// metadata about the graph structure and execution flow.
     ///
     /// Returns a `LuminairPie` containing all the trace tables and execution resources.
-    fn gen_trace(&mut self, settings: &mut CircuitSettings) -> Result<LuminairPie, TraceError> {
+    fn gen_trace(&mut self, settings: &mut CircuitSettings) -> Result<LuminairPie, LuminairError> {
         // Track the number of views pointing to each tensor so we know when to clear
         if self.linearized_graph.is_none() {
             self.toposort();
@@ -496,195 +437,6 @@ impl LuminairGraph for Graph {
                 op_counter,
                 max_log_size,
             },
-        })
-    }
-
-    /// Generates a STWO proof for the computation graph execution.
-    ///
-    /// Takes the `LuminairPie` (containing execution traces) and `CircuitSettings`.
-    /// It orchestrates the STWO proving protocol:
-    /// 1. Sets up the prover, channel, and commitment scheme.
-    /// 2. Commits to the preprocessed trace.
-    /// 3. Commits to the main execution trace components (add, mul, sin, etc.).
-    /// 4. Commits to the interaction trace.
-    /// 5. Executes the Stwo prover.
-    /// Returns a `LuminairProof` containing the claims and the STARK proof.
-    fn prove(
-        &mut self,
-        pie: LuminairPie,
-        settings: CircuitSettings,
-    ) -> Result<LuminairProof<Blake2sMerkleHasher>, LuminairError> {
-        // ┌──────────────────────────┐
-        // │     Protocol Setup       │
-        // └──────────────────────────┘
-        tracing::info!("Protocol Setup");
-        let config: PcsConfig = PcsConfig::default();
-        let max_log_size = pie.execution_resources.max_log_size;
-        let twiddles = SimdBackend::precompute_twiddles(
-            CanonicCoset::new(max_log_size + config.fri_config.log_blowup_factor + 2)
-                .circle_domain()
-                .half_coset,
-        );
-        // Setup protocol.
-        let channel = &mut Blake2sChannel::default();
-        let mut commitment_scheme =
-            CommitmentSchemeProver::<_, Blake2sMerkleChannel>::new(config, &twiddles);
-
-        // ┌───────────────────────────────────────────────┐
-        // │   Interaction Phase 0 - Preprocessed Trace    │
-        // └───────────────────────────────────────────────┘
-
-        tracing::info!("Preprocessed Trace");
-        // Convert lookups in circuit settings to preprocessed column.
-        let lut_cols = lookups_to_preprocessed_column(&settings.lookups);
-        let preprocessed_trace = PreProcessedTrace::new(lut_cols);
-        let mut tree_builder = commitment_scheme.tree_builder();
-        tree_builder.extend_evals(preprocessed_trace.gen_trace());
-        // Commit the preprocessed trace
-        tree_builder.commit(channel);
-
-        // ┌───────────────────────────────────────┐
-        // │    Interaction Phase 1 - Main Trace   │
-        // └───────────────────────────────────────┘
-
-        tracing::info!("Main Trace");
-        let mut main_claim = LuminairClaim::default();
-        let mut interaction_claim_gen = LuminairInteractionClaimGenerator::default();
-        let mut tree_builder = commitment_scheme.tree_builder();
-
-        for table in pie.trace_tables.clone() {
-            match table {
-                TraceTable::Add { table } => {
-                    let claim_gen = add::witness::ClaimGenerator::new(table);
-                    let (cl, in_cl_gen) = claim_gen.write_trace(&mut tree_builder)?;
-                    main_claim.add = Some(cl.clone());
-                    interaction_claim_gen.add = Some(in_cl_gen);
-                }
-                TraceTable::Mul { table } => {
-                    let claim_gen = mul::witness::ClaimGenerator::new(table);
-                    let (cl, in_cl_gen) = claim_gen.write_trace(&mut tree_builder)?;
-                    main_claim.mul = Some(cl.clone());
-                    interaction_claim_gen.mul = Some(in_cl_gen);
-                }
-                TraceTable::Recip { table } => {
-                    let claim_gen = recip::witness::ClaimGenerator::new(table);
-                    let (cl, in_cl_gen) = claim_gen.write_trace(&mut tree_builder)?;
-                    main_claim.recip = Some(cl.clone());
-                    interaction_claim_gen.recip = Some(in_cl_gen);
-                }
-                TraceTable::Sin { table } => {
-                    let claim_gen = sin::witness::ClaimGenerator::new(table);
-                    let (cl, in_cl_gen) = claim_gen.write_trace(&mut tree_builder)?;
-                    main_claim.sin = Some(cl.clone());
-                    interaction_claim_gen.sin = Some(in_cl_gen);
-                }
-                TraceTable::SinLookup { table } => {
-                    let claim_gen = lookups::sin::witness::ClaimGenerator::new(table);
-                    let (cl, in_cl_gen) = claim_gen.write_trace(&mut tree_builder)?;
-                    main_claim.sin_lookup = Some(cl.clone());
-                    interaction_claim_gen.sin_lookup = Some(in_cl_gen);
-                }
-                TraceTable::SumReduce { table } => {
-                    let claim_gen = sum_reduce::witness::ClaimGenerator::new(table);
-                    let (cl, in_cl_gen) = claim_gen.write_trace(&mut tree_builder)?;
-                    main_claim.sum_reduce = Some(cl.clone());
-                    interaction_claim_gen.sum_reduce = Some(in_cl_gen);
-                }
-                TraceTable::MaxReduce { table } => {
-                    let claim_gen = max_reduce::witness::ClaimGenerator::new(table);
-                    let (cl, in_cl_gen) = claim_gen.write_trace(&mut tree_builder)?;
-                    main_claim.max_reduce = Some(cl.clone());
-                    interaction_claim_gen.max_reduce = Some(in_cl_gen);
-                }
-                TraceTable::Sqrt { table } => {
-                    let claim_gen = sqrt::witness::ClaimGenerator::new(table);
-                    let (cl, in_cl_gen) = claim_gen.write_trace(&mut tree_builder)?;
-                    main_claim.sqrt = Some(cl.clone());
-                    interaction_claim_gen.sqrt = Some(in_cl_gen);
-                }
-            }
-        }
-        // Mix the claim into the Fiat-Shamir channel.
-        main_claim.mix_into(channel);
-        // Commit the main trace.
-        tree_builder.commit(channel);
-
-        // ┌───────────────────────────────────────────────┐
-        // │    Interaction Phase 2 - Interaction Trace    │
-        // └───────────────────────────────────────────────┘
-
-        tracing::info!("Interaction Trace");
-        let interaction_elements = LuminairInteractionElements::draw(channel);
-        let mut interaction_claim = LuminairInteractionClaim::default();
-        let mut tree_builder = commitment_scheme.tree_builder();
-        let node_elements = &interaction_elements.node_elements;
-        let lookup_elements = &interaction_elements.lookup_elements;
-        if let Some(claim_gen) = interaction_claim_gen.add {
-            let claim = claim_gen.write_interaction_trace(&mut tree_builder, node_elements);
-            interaction_claim.add = Some(claim)
-        }
-        if let Some(claim_gen) = interaction_claim_gen.mul {
-            let claim = claim_gen.write_interaction_trace(&mut tree_builder, node_elements);
-            interaction_claim.mul = Some(claim)
-        }
-        if let Some(claim_gen) = interaction_claim_gen.recip {
-            let claim = claim_gen.write_interaction_trace(&mut tree_builder, node_elements);
-            interaction_claim.recip = Some(claim)
-        }
-        if let Some(claim_gen) = interaction_claim_gen.sin {
-            let claim = claim_gen.write_interaction_trace(
-                &mut tree_builder,
-                node_elements,
-                &lookup_elements.sin,
-            );
-            interaction_claim.sin = Some(claim)
-        }
-        if let Some(claim_gen) = interaction_claim_gen.sin_lookup {
-            let mut sin_luts = preprocessed_trace.columns_of::<SinPreProcessed>();
-            sin_luts.sort_by_key(|c| c.col_index);
-
-            let claim = claim_gen.write_interaction_trace(
-                &mut tree_builder,
-                &lookup_elements.sin,
-                &sin_luts,
-            );
-            interaction_claim.sin_lookup = Some(claim)
-        }
-        if let Some(claim_gen) = interaction_claim_gen.sum_reduce {
-            let claim = claim_gen.write_interaction_trace(&mut tree_builder, node_elements);
-            interaction_claim.sum_reduce = Some(claim)
-        }
-        if let Some(claim_gen) = interaction_claim_gen.max_reduce {
-            let claim = claim_gen.write_interaction_trace(&mut tree_builder, node_elements);
-            interaction_claim.max_reduce = Some(claim)
-        }
-        if let Some(claim_gen) = interaction_claim_gen.sqrt {
-            let claim = claim_gen.write_interaction_trace(&mut tree_builder, node_elements);
-            interaction_claim.sqrt = Some(claim)
-        }
-        // Mix the interaction claim into the Fiat-Shamir channel.
-        interaction_claim.mix_into(channel);
-        // Commit the interaction trace.
-        tree_builder.commit(channel);
-
-        // ┌──────────────────────────┐
-        // │     Proof Generation     │
-        // └──────────────────────────┘
-        tracing::info!("Proof Generation");
-        let component_builder = LuminairComponents::new(
-            &main_claim,
-            &interaction_elements,
-            &interaction_claim,
-            &preprocessed_trace,
-            &settings.lookups,
-        );
-        let components = component_builder.provers();
-        let proof = prover::prove::<SimdBackend, _>(&components, channel, commitment_scheme)?;
-
-        Ok(LuminairProof {
-            claim: main_claim,
-            interaction_claim,
-            proof,
         })
     }
