+++ crates/graph/src/tests/ops.rs
@@ -2,6 +2,7 @@ use super::{assert_close, random_vec_rng};
 use crate::graph::LuminairGraph;
 use crate::StwoCompiler;
 use crate::{binary_test, unary_test};
+use luminair_prover::prover::prove;
 use luminal::prelude::*;
 use luminal_cpu::CPUCompiler;
 use rand::{rngs::StdRng, SeedableRng};
@@ -46,9 +47,7 @@ fn test_sum_reduce() {
     let trace = cx
         .gen_trace(&mut settings)
         .expect("Trace generation failed");
-    let proof = cx
-        .prove(trace, settings.clone())
-        .expect("Proof generation failed");
+    let proof = prove(trace, settings.clone()).expect("Proof generation failed");
     cx.verify(proof, settings)
         .expect("Proof verification failed");
 
@@ -94,9 +93,7 @@ fn test_max_reduce() {
     let trace = cx
         .gen_trace(&mut settings)
         .expect("Trace generation failed");
-    let proof = cx
-        .prove(trace, settings.clone())
-        .expect("Proof generation failed");
+    let proof = prove(trace, settings.clone()).expect("Proof generation failed");
     cx.verify(proof, settings)
         .expect("Proof verification failed");
