+++ crates/utils/Cargo.toml
@@ -1,12 +1,11 @@
 [package]
-name = "simple"
+name = "luminair_utils"
 version.workspace = true
 edition.workspace = true
 repository.workspace = true
 license.workspace = true
 
 [dependencies]
+thiserror.workspace = true
+luminair_air = { path = "../air" }
 stwo-prover.workspace = true
-luminal.workspace = true
-luminair_graph = { path = "../../crates/graph" }
-luminair_prover = { path = "../../crates/prover" }
